Move pull into its own file
Towards #290
